Create access helper - Enables Access behaviour on using module.
