+++ lib/aurora_uix/layout/build_tree.ex
@@ -1,0 +1,4 @@
+defmodule Aurora.Uix.Layout.BuildTree do
+  def build_tree(_resource_metadata, _layout_path) do
+  end
+end
