+++ lib/aurora_uix/tree_path.ex
@@ -1,0 +1,4 @@
+defmodule Aurora.Uix.TreePath do
+  use Aurora.Uix.AccessHelper
+  defstruct [:name, :tag, config: [], opts: [], inner_elements: []]
+end
